Fix failing crontab job.  Simplify email secrets handling.
After adding ~/Dropbox/proj/dotfiles/bin to the path, the
new_oist_careers invocation in my .crontab was failing, because it was
hardcoded to ~/bin dir.

Removed the dummy .forward from dotfiles and replaced it with a real
.forward in my secrets.

Modified dotf to not replace dummy email (foo@example.com) with my real
email in .crontab or .forward.

Moved .crontab from copy/ to link_home/, so I can modify ~/.crontab when
I need to, instead of (laboriously) modifying
~/Dropbox/proj/dotfiles/copy/.crontab and redeploying it to ~/.  This
avoids the error of modifying ~/.crontab too.